Change base image to one that supports more archs.
With the new build image python and the main image one can now build
polemarch on following architectures(supported by both images)
 - 386
 - amd64
 - arm/v5
 - arm/v7
 - arm64/v8
 - mips64le
 - ppc64le
 - s390x

The usecase for me personally was to build to container on a RPI4 and
run it there. Altough alpine supports more architectures,
this was not possible before as the used Tux builder image is x86
only and also has no puplic available dockerfile to rebuild it for
aarch64. When changing this to python:3.6-buster one also has no need to
maintain the tux image anymore if it is only made to build polemarch of
course.